refactor to new dir logic and add Initialize to a menu point
