+++ src/logic/convert.py
@@ -1,0 +1,8 @@
+def join_array(array: list[str]) -> str:
+    """
+    Join text.
+
+    :param array:
+    :return:
+    """
+    return " ".join(array)
