+++ src/main.py
@@ -1,6 +1,7 @@
 import tkinter
 from typing import Type
 
+from src.logic.initialize_database import add_special_psa
 from src.views.equipmentgui import EquipmentGUI
 from src.views.membergui import MemberGUI
 from src.views.psagui import PsaGUI
@@ -23,7 +24,7 @@ class MainGUI(tkinter.Tk):
     def __init__(self, *args, **kwargs) -> None:  # type: ignore
         tkinter.Tk.__init__(self, *args, **kwargs)
 
-        self.title("ffwgerätewart")
+        self.title("ffw-geraetewart")
         # self.geometry("200x200")
 
         menubar = tkinter.Menu(self)
@@ -40,9 +41,13 @@ class MainGUI(tkinter.Tk):
         equipment_menu = tkinter.Menu(menubar)
         equipment_menu.add_command(label="Anzeigen", command=lambda: self.show_frame("EquipmentGUI"))
 
+        about_menu = tkinter.Menu(menubar)
+        about_menu.add_command(label="Initialisieren Spezieller PSA Vorlagen", command=lambda: add_special_psa())
+
         menubar.add_cascade(label="Mitglieder", menu=member_menu, underline=0)
         menubar.add_cascade(label="PSA", menu=psa_menu, underline=0)
         menubar.add_cascade(label="Equipment", menu=equipment_menu, underline=0)
+        menubar.add_cascade(label="Über", menu=about_menu, underline=0)
 
         # the container is where we'll stack a bunch of frames
         # on top of each other, then the one we want visible
